basic functionality
can set a pin as input output
can read inputs
can set output as high or low

it appears to be working well for pins 1 through 8!

should probably think about making the silk screen and the coded pins match, ie there is no pin 8 but pin 7 and there is a pin0 in silk but not a pin0 in the code.